Adding admin functionality - list out all stored data

+++ src/main/java/com/mark/model/dal/ApplicationState.java
@@ -7,13 +7,13 @@ import org.joda.time.DateTime;
 public class ApplicationState implements Serializable {
 
 	private static final long serialVersionUID = 1L;
-	private Integer flightApiCount;
+	private Long flightApiCount;
 	private DateTime date;
 	
-	public Integer getFlightApiCount() {
+	public Long getFlightApiCount() {
 		return flightApiCount;
 	}
-	public void setFlightApiCount(Integer flightApiCount) {
+	public void setFlightApiCount(Long flightApiCount) {
 		this.flightApiCount = flightApiCount;
 	}
 	public DateTime getDate() {
